React DOM. Separate files for external declarations
