+++ kotlin-react-dom/src/main/kotlin/react/dom/render.kt
@@ -1,0 +1,14 @@
+@file:JsModule("react-dom")
+@file:JsNonModule
+
+package react.dom
+
+import org.w3c.dom.Element
+
+// See https://reactjs.org/docs/react-dom.html
+
+external fun render(
+    element: dynamic,
+    container: Element?,
+    callback: () -> Unit = definedExternally,
+)
