+++ kotlin-react-dom/src/main/kotlin/react/dom/unmountComponentAtNode.kt
@@ -1,0 +1,12 @@
+@file:JsModule("react-dom")
+@file:JsNonModule
+
+package react.dom
+
+import org.w3c.dom.Element
+
+// See https://reactjs.org/docs/react-dom.html
+
+external fun unmountComponentAtNode(
+    domContainerNode: Element?,
+)
